First steps toward implementing dropdowns for choosing a word and example
Work in progress. The handleWordSelect method successfully sets the selectedWord key in the state object to the appropriate word, but the handleExampleSelect method reveals a deeper issue with nested objects in a state object. Currently throws an error if trying to choose an example for the selected word (see line 117). The selectedWord key in the state object is set to an empty string upon loading, so we implemented a [temporary] solution for not rendering the dropdown menu for examples unless the user selects a word.

Co-Authored-By: Pwnybaloney <36424301+Pwnybaloney@users.noreply.github.com>